tests/ublox_control/requirements-test.txt
# GRPC Python setup requirements
coverage>=4.0
cython >=3.1.1
protobuf>=6.30.0,<7.0.0
wheel>=0.29

grpcio>=1.73.0
grpcio_reflection>=1.73.0
grpcio_tools>=1.73.0

# async libraries
#watchfiles>=1.1.0
#aiofiles>=24.1.0

# observability
json5>=0.12.0
rich>=14.0.0
redis>=3.5.3

# ublox control
pyserial>=3.5
pyubx2>=1.2.43
redis>=3.5.3

# data analysis
#seaborn>=0.13.2
#matplotlib>=3.8.4
pandas>=2.3.0
numpy>=1.26.4
#jupyter>=1.1.1

# panoseti utils
netifaces >= 0.11.0
psutil >= 7.0.0

# testing
pytest >= 8.4.1
pytest-asyncio >= 1.1.0
